requirements.txt
CacheControl==0.12.6
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
Django==2.2
django-background-tasks==1.2.5
django-compat==1.0.15
django-pusher==0.1.1
firebase-admin==4.0.0
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
httplib2==0.17.0
idna==2.9
msgpack==1.0.0
ndg-httpsclient==0.5.1
protobuf==3.11.3
psycopg2==2.8.4
pusher==2.1.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyNaCl==1.3.0
pyOpenSSL==19.1.0
python-firebase==1.2
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
